Changes for configs and moving some hard-coded IP values (#7)
* Changes:
-Splitting out separate applications for the World and Channel servers based off the existing Lobby server.
-Adding sync/async option for TcpConnection since the two new servers shouldn't even start if they cannot connect to one of the other servers.

Issues:
-*Connection classes act as both client and server connections. The ConnectionSuccessful function needs to act differently for clients and servers
-Servers need to handle multiple ports better to support multiple world and channel servers

* Changes:
-Civet dropped for comp_channel and comp_world, leaving the handlers as place holders for the time being
-Removing the generated ResourceLogin.c files for comp_channel and comp_world
-Replacing tabs with spaces

* Changes:
-Removing unused Handlers and explicit connections for world and channel servers
-Adding InternalServer, InternalServerWorker and InternalConnection for use by non-lobby servers
-Adding worker thread concept to internal servers
-Fixed GCC compiler issue

Issues:
-Copied some code that should probably be consolidated from Lobby
-Although message queuing is working fine between the three servers, it may not be set up properly once we start to implement real packet handling

* [FIX] Lobby connection bugs.

* [FIX] Combine commands into a single outgoing packet before encryption (and compression in the channel server case).

* [FIX] Messages.

* [FIX] Race condition.

* [ADD] Ensure the lobby connection is encrypted.

* [ADD] Lobby testing configuration.

* [ADD] Initial worker structure.

* [FIX] Bugs

* [FIX] Missing include.

* Changes:
-Implementing libcomp::Worker into InternalServer
-Stubbing out logic points for message handling for the world and channel servers
-Removing InternalServerWorker

* Simplifying server code to be more in line with LobbyServer implementation and adding messages to the client server connection threads to fix a race condition discovered by Omega

* Removing GetMessageHandler from WorldServer and re-copying functions from LobbyConnection to InternalConnection (these two still need to have code moved to a base class).

* Objgen base object support and World/Channel server configs

Changes===
-Adding base object support to objgen
-Adding WorldConfig and ChannelConfig generated classes and consolidated the LobbyConfig DH pair code
-Fixed issue with not setting the message queue for InternalConnection while connecting to a host server

* Backing out default config file path for non-Windows

* Fixing last deploy :/

* More useful configs and lobby connection bug fix on Windows

Changes:
-Fixed a run-time bug stopping the lobby from accepting game client packets
-Moved all IP and port information into the server configs
-Server configs have been moved out of libcomp into their respective projects
-BaseServer constructor now requires a ServerConfig and path to configure instead of raw address/port info
-comp_lobby, comp_world and comp_channel now take an optional parameter to point to an explicit config path

* Moving ServerConfig and the handling functions into BaseServer from TcpServer

* Moving data members for objgen back into private accessors and modifying derived classes accordingly

* Removing a ChannelServer function I missed in the last commit and ignoring the server/lobby/res so I don't accidentally push it

* Dropping InternalServer since it was removed in the main repo and isn't used in this fork

* Some cleanup for the merge

+++ libcomp/src/TcpConnection.h
@@ -73,7 +73,7 @@ public:
     static std::vector<char> GenerateDiffieHellmanSharedData(
         DH *pDiffieHellman, const String& otherPublic);
 
-    bool Connect(const String& host, int port = 0, bool async = true);
+    bool Connect(const String& host, uint16_t port = 0, bool async = true);
 
     void Close();
